refactor: create a hook useEntryDataForm
